fatal error fixed!!!!
you can find the difference in file "seelctio.cpp line 260"

we add "if (total_size < len)
 sock_data  = getSOCKETPacket(s);" there.

because, when a packet contain some information that belonged to two httppacket, first, the packet is finished, and we remove it from sockets_map_, then we should deal the rest data, actually we need a new httppacket. but we always use the last one.