issue 57 task
if the DNS is 'http://code.google.com/p/httpcontentparser/issues/entry', we
cannot get the main name of DNS by the function, just because we must pass
an DNS name like 'code.google.com' but not 'code.google.com/a/b/c'. 

we suppose that user will specified an host_name, but not the string like
the one in browse address item. we should enhance the function.