I found a bug: 1. enter parent model, and set Entertime and EyecareTime, 2. switch to the child mode.
but I found that when I call the getRemainTime(), the result is not changed.

Perhaps the function mentioned in  Issue 43  are not completely implemented.



just because in the interface IAppSetting 'IAppSetting', when the application switch model, we check the time's stopped or not. 

I write like this:
    if (false == isTimer stopped) {
        restart();
    }

Actually, we should write it 
I write like this:
    if (true == isTimer stopped) {
        restart();
    }